requirements.txt
requests
esdk-obs-python
minio
oss2
